Relocate all steps to appropriate command folders
