+++ src/commands/farewell/farewellConfirmation.js
