+++ src/commands/farewell/farewellRejection.js
