+++ src/commands/tokenAdd/addCW20.js
