+++ src/commands/tokenAdd/addCW721.js
