+++ src/commands/tokenAdd/addNativeToken.js
