+++ src/commands/tokenAdd/handleCW20Entry.js
