+++ src/commands/tokenAdd/hasCW20.js
