+++ src/commands/tokenAdd/hasCW721.js
