+++ src/commands/tokenAdd/nativeTokenSTARS.js
