+++ src/commands/tokenAdd/nativeTokenSuggestion.js
