+++ src/commands/tokenAdd/needsCW20.js
