+++ src/commands/tokenAdd/promptTokenAmount.js
