+++ src/commands/tokenAdd/promptTokenName.js
