+++ src/commands/tokenAdd/stargaze.js
