+++ src/commands/tokenEdit/editCheck.js
