+++ src/commands/tokenEdit/editRoleAmount.js
