+++ src/commands/tokenEdit/handleRoleAmountEdit.js
