+++ src/commands/tokenEdit/handleRoleNameEdit.js
