+++ src/commands/tokenRemove/removeConfirmation.js
