+++ src/commands/tokenRemove/removeVerify.js
